+++ Editor/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("GitPackageManager.Editor.Tests")]
+[assembly: InternalsVisibleTo("Essentials.GitPackageManager.Editor.Tests")]