+++ Editor/GitPackageBootstrap.cs
@@ -2,7 +2,7 @@ using System.IO;
 using UnityEditor;
 using UnityEngine;
 
-namespace GitPackageManager.Editor
+namespace Essentials.GitPackageManager.Editor
 {
     [InitializeOnLoad]
     internal static class GitPackageBootstrap