+++ Editor/PackageManagerWindow.ListPane.cs
@@ -2,7 +2,7 @@ using System;
 using UnityEditor;
 using UnityEngine;
 
-namespace GitPackageManager.Editor
+namespace Essentials.GitPackageManager.Editor
 {
     public partial class GitPackageManagerWindow
     {