+++ Editor/PackageManagerWindow.Styles.cs
@@ -1,7 +1,7 @@
 using UnityEditor;
 using UnityEngine;
 
-namespace GitPackageManager.Editor
+namespace Essentials.GitPackageManager.Editor
 {
     public partial class GitPackageManagerWindow
     {